# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Gorax - a powerful workflow automation platform with webhooks, scheduling, and real-time execution monitoring.\n\n## Authentication\nAll API endpoints (except /health, /ready, and webhook triggers) require authentication.\nIn development mode, use the X-User-ID header. In production, use Ory Kratos session cookies.\n\n## Multi-tenancy\nAll authenticated endpoints require a valid X-Tenant-ID header to identify the tenant context.\n\n## Rate Limiting\nAPI requests are rate-limited based on tenant quotas. Check response headers for limit information.\n\n## Error Handling\nErrors follow a consistent format with appropriate HTTP status codes and error messages.",
        "title": "Gorax Workflow Automation API",
        "contact": {
            "name": "Gorax Support",
            "url": "https://github.com/gorax/gorax",
            "email": "support@gorax.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/webhooks": {
            "get": {
                "security": [
                    {
                        "TenantID": []
                    },
                    {
                        "UserID": []
                    }
                ],
                "description": "Returns a paginated list of webhooks for the authenticated tenant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhooks"
                ],
                "summary": "List webhooks",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Maximum number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of webhooks",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "TenantID": []
                    },
                    {
                        "UserID": []
                    }
                ],
                "description": "Creates a new webhook endpoint for a workflow",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhooks"
                ],
                "summary": "Create webhook",
                "parameters": [
                    {
                        "description": "Webhook creation request",
                        "name": "webhook",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.CreateWebhookRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created webhook",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request or validation error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/webhooks/{id}": {
            "get": {
                "security": [
                    {
                        "TenantID": []
                    },
                    {
                        "UserID": []
                    }
                ],
                "description": "Retrieves a webhook by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhooks"
                ],
                "summary": "Get webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Webhook ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Webhook details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Webhook not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/webhooks/{id}/regenerate-secret": {
            "post": {
                "security": [
                    {
                        "TenantID": []
                    },
                    {
                        "UserID": []
                    }
                ],
                "description": "Regenerates the secret key for webhook signature verification",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhooks"
                ],
                "summary": "Regenerate webhook secret",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Webhook ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Webhook with new secret",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Webhook not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api_handlers.CreateWebhookRequest": {
            "type": "object",
            "required": [
                "name",
                "path",
                "workflowId"
            ],
            "properties": {
                "authType": {
                    "type": "string",
                    "enum": [
                        "none",
                        "signature",
                        "basic",
                        "api_key"
                    ]
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer",
                    "maximum": 3,
                    "minimum": 0
                },
                "workflowId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "SessionCookie": {
            "description": "Ory Kratos session cookie (production mode)",
            "type": "apiKey",
            "name": "ory_kratos_session",
            "in": "cookie"
        },
        "TenantID": {
            "description": "Tenant identifier for multi-tenant isolation",
            "type": "apiKey",
            "name": "X-Tenant-ID",
            "in": "header"
        },
        "UserID": {
            "description": "User identifier (development mode only)",
            "type": "apiKey",
            "name": "X-User-ID",
            "in": "header"
        }
    },
...
}